Move some generic util functions from vacuumcontainers to utils module
